jingshi/requirements.txt
opencv-python
Pillow
pygetwindow
pyautogui
numpy
pyinstaller
